Changes to interact with POJOPlugin through the loaded JMI Model.  What ends up happening now is the JMI is annotated with the missing details of the model by the specific plugin.  In this manner, the set of cartridges can be looked at more like a stream than simply independent entities that are all just sharing a dataset.
One of the things that this made me realize the need for "layering", a concept in JMI that is similar to Photoshop.  We can add layers into the JMI such that changes could be made on disposable layers, allowing for various clients to see data views with or without the effects of changes like we are now doing.

Also adding the full source for the generated UML interfaces, since the CWM interfaces are already in and it's nice to have the symbolics for the method calls when developing.